It's LOADING RIGHT????
Hey guys, I've spent many hours...into early morning to try and get this
map working for us? and guess what, I've done it! Just have some minor
adjustments to do which will take a period of 5 mins and it should be
all good for us to go on and complete this assignment!